Fix check registration status to be robust against network failures (#3713)
_legacy_apy_registration_check checks if the system is registered or not.
The server can response:
A 200 status, telling the system is registered. Unless if the "unregistered_at"
parameter is not null. In that case the system is unregistered.
A 404 status with a json telling the system is not registered.
Other status error, telling the HTTPS connection faced some problems.

This commit adds a condition to only check the json response if the status code
is 200 or 404 which means the server has information about the registration status
of the system.

Resolves: ESSNTL-4444

Signed-off-by: ahitacat <ahitacat@redhat.com>
(cherry picked from commit f6085706cecfcfc68e4d3436a2c3a26239a3b0fe)