refactor(client): remove hacky proxy code (#3730)
Requests and urllib3 now support proxy authentication natively so we
don't need this proxy hack stuff anymore. Instead we can just directly
read in the proxy authentication.

The auth split from the url is left in so that debu messages can still
work without printing the authentication to the console.

Signed-off-by: Stephen Adams <stephen.adams@redhat.com>