fix: resolve missing insights-client.ppid error
This patch attempts to resolve ABRT core dumps
caused by attempting to remove the insights-client.ppid
file during program exit.

The 'write_to_disk' method is modified to
include a redundant try-except to catch
any race conditions that may cause this rare error
and a test case is included to mock this behaviour.

Signed-off-by: Jason Jerome <jajerome@redhat.com>