fix: resolve missing 'insights-client/lib' error in playbook verification tests
This fix creates improvements to existing
ansible playbook verification tests by
creating temp directories in tests that
need a gpnughome directory.

The 'vars' verification tests are updated
to mock the 'getRevocationList' method
since it's not needed for the purpose of
these tests. The other tests that use
code related to public keys use temp
directories and mock the 'insights_core_lib_dir'.

Signed-off-by: Jason Jerome <jajerome@redhat.com>