RHEL-2480: Do not create /root/.gnupg/ directory by accident (#3930)
* feat(pgp): Add crypto.py with PGP signature verification

* Card ID: CCT-131
* Card ID: RHEL-2480
* Card ID: RHEL-2482

This patch adds a self-contained and isolated GPG verification
environment. It runs GPG in an isolated environment where only selected
PGP keys are allowed to check the file signature matches its file.

GPG creates a directory `$HOME/.gnupg/` every time it performs some
operation. When run under root, but not manually (e.g. via
subscription-manager Cockpit plugin), it tries to create and write to
this directory, which pollutes user directories and/or causes SELinux
denials.

This patch utilizes the `--homedir` argument GPG supports in order to
move the GPG home directory to a temporary directory for the time of the
transaction. After the GPG action is performed, the directory is cleaned
up.

Signed-off-by: mhorky <mhorky@redhat.com>

* feat(pgp): Use crypto.py during Egg and Collection verification

* Card ID: CCT-131
* Card ID: RHEL-2480
* Card ID: RHEL-2482

Signed-off-by: mhorky <mhorky@redhat.com>

* chore: Add .gitleaks.toml

This configuration file ensures the dummy PGP public/private key pair
is not flagged by security tools as a false positive.

Signed-off-by: mhorky <mhorky@redhat.com>

---------

Signed-off-by: mhorky <mhorky@redhat.com>