fix: check status created a machine-id file (#3965)
* feat: add machine_id_exists method

This method returns a boolean whether the machine-id file exists or
not. This will be used to check the file before call, generate_machine_id
to get the machine-id, as this last method will create a new id if it
doesn't exist.

Resolves: CCT-161

Signed-off-by: ahitacat <ahitacat@redhat.com>

* fix: use machine_id_exists before checking host status

When checking registration status, first check if the machine-id file
exists. When checking status it is not needed to generate a machine-id
file if it doesn't exists.

Resolves: CCT-161

Signed-off-by: ahitacat <ahitacat@redhat.com>

* fix: Change test that generated id and now use get_machineid

Also create a new test to check a new machine id is created if the
file does not exists.

Signed-off-by: ahitacat <ahitacat@redhat.com>

---------

Signed-off-by: ahitacat <ahitacat@redhat.com>
(cherry picked from commit 5697ed4b35fd4edde1eb8c2e8d619a764364d648)