feat: Reuse subscription-manager identity for machine-id (#4057)
* feat: Add logging into utilities.write_to_disk

Sometimes it is hard to see which files are changed. This should make it
easier to investigate.

Signed-off-by: mhorky <mhorky@redhat.com>

* feat: Reuse subscription-manager identity for machine-id

* Card ID: CCT-161

insights-client sometimes incorrectly regenerates identity while it
shouldn't. Because the machine-id file is created preemptively, instead
of at registration time (and deleted if the registration fails or
Inventory returns error status signaling the system has been deleted in
there), this is the easiest way to ensure the UUID will stay
semi-stable.

subscription-manager already provides a stable way to identify a system.
To ensure external tooling doesn't get confused, we should try to keep
them consistent.

Signed-off-by: mhorky <mhorky@redhat.com>

---------

Signed-off-by: mhorky <mhorky@redhat.com>