feat: Improve debugging experience (#4094)
To improve client-side debugging, made the core log its version along its path whenever verbose mode is enabled. It is no longer necessary to pass --debug, which was the sole use case of this undocumented flag. Like this, the core version can be found in any log file captured from a customer’s machine.

The core path is now less confusing. It points to core package top directory, not to a module containing the log call. The logic is the same as in the Client.

To find the exact place where logged events occur, a line number is now part of every log message.

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>
(cherry picked from commit 47db172e9fae745792417be5160c654f7aa4c697)