fix(test): Catch errors caused by GPG
There exists a race condition when removing files in a directory in
which GPG has saved its runtime data.

When GPG detects the files are getting removed, it will kill its socket
and start deleting the files. In that case, `rmtree` may try to delete a
file that has already been deleted by GPG, causing `OSError`.

Signed-off-by: mhorky <mhorky@redhat.com>