chore: Disable legacy upload
* Card ID: CCT-252
* Card ID: RHINENG-7569

This patch disables the legacy APIs. The code and configuration option
will be removed later; this is the first step to get rid of the
duplication and differences.

Only the absolutely necessary changes were made to tests; without these
removals the suite wouldn't pass. The rest will be cleaned up later.

Signed-off-by: mhorky <mhorky@redhat.com>