fix: Prevent test teardown issues of fast_fetch
Two tests were mocking result of tempfile to prevent files from being
written to filesystem during a test. A few years ago, a cleanup method
was added to the tested code, to ensure the temporary directory the egg
was downloaded to is cleaned up. Since then, these tests have been
emitting warnings that Mock cannot be `lstat`ed.

This patch fixes it.

Signed-off-by: mhorky <mhorky@redhat.com>