Fix skipped assert when `expected` is false
`run_test` ignores `expected` if its value does not evaluate to a true
condition. This means that if, for example, the expected result from a
component is an empty list or the actual value `None`, no assertion is
performed leading to undetected test failures.

To fix this, the default value for the `expected` parameter is a
sentinel object that allows the detection of an expected value being
passed in regardless if it is `None` or any other condition that does
not evaluate to true.

Signed-off-by: Larry O'Leary <loleary@redhat.com>