Fix skipped assertions for itests (#4207)
* Fix skipped assert when `expected` is false

`run_test` ignores `expected` if its value does not evaluate to a true
condition. This means that if, for example, the expected result from a
component is an empty list or the actual value `None`, no assertion is
performed leading to undetected test failures.

To fix this, the default value for the `expected` parameter is a
sentinel object that allows the detection of an expected value being
passed in regardless if it is `None` or any other condition that does
not evaluate to true.

Signed-off-by: Larry O'Leary <loleary@redhat.com>

* Remove expected value from integration tests

Assertion is performed by the test itself. The epxected value of `None`
is invalid for these tests. Remove `expected` argument from `run_test`
call.

Signed-off-by: Larry O'Leary <loleary@redhat.com>
(cherry picked from commit ac4f45e442b2533398eab56f7bb1640605696bd0)