feat(client): write .last-upload.results also after non-legacy uploads
While that file is deprecated and it is good that it is not written
after non-legacy uploads, in practice it is actually used by tooling
that uses insights-client.

To avoid breaking functionalities in other software when switching to
non-legacy, write /etc/insights-client/.last-upload.results also after
non-legacy uploads. While those softwares are fixed to not rely on that
file, at least this is one small step ahead in switching the default
away to non-legacy without regressions elsewhere.

Implements CCT-710

Signed-off-by: Pino Toscano <ptoscano@redhat.com>