Catch Timeout on test urls (#4248)
The internal methods used by the --test-connection command only used to catch ConnectionError, including ConnectionTimeout, but not other timeouts such as ReadTimeoutError. As a result, in case of a read timeout,  the failure is not fully handled. That means:

* an error message containing the failed URL is not logged;
* and in case of legacy upload, the URL fallback mechanism is interrupted.

Re-used the existing REQUEST_FAILED_EXCEPTIONS list to process all timeouts in the same way.

Card IDs:

* CCT-859

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>