Update with some speed improvements
Xiangce Liu suggested a few improvements; the main one was to compile
the set of keywords only from the first row if we know that the keys
don't change between rows.  In the few cases they do, the caller needs
to set the `row_keys_change` option to `True`.  This then simplified the
logic of transforming the keys, which meant that we only have to do a
'forward' translation rather than reverse.

Signed-off-by: Paul Wayper <paulway@redhat.com>