fix(client): checkin for unregistered hosts should fail (#4274)
* Card ID: CCT-891
* Card ID: RHINENG-13984

The `insights-client --checkin` command currently succeeds on hosts
that are not registered with `insights-client`, but are registered
with `subscription-manager`. This fix addresses the issue to ensure
the `--checkin` command properly fails as expected when the client is
not registered.

Signed-off-by: pkoprda <pkoprda@redhat.com>
(cherry picked from commit b86efb5bed16d92d1acad3dfc240f520f1fc8e6d)