fix(ci): Use dynamic repo URL for python26-test (#4314)
The harcoded RedHatInsights URL makes the Python 2.6 tests fail when a pull request is created in a fork. Using a dynamic URLs makes the GitHub Action to clone from the fork and use the code in the pull request.

github.repositoryUrl variable can’t be used, because cloning using  git:// protocol fails. HTTPS url is thus composed instead.

Card IDs:

* RHINENG-14956

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>
Signed-off-by: max <mclinton@redhat.com>