fix: warnings keywords over-matching in LvmFullReport (#4391)
The former added warning keywords filtering breaks one of the testcase
in extraction-rules repo, eg. "duplicate" is a keyword from the lvm
report json, and should not be filtered out as warning messages.

Consider all the keywords in find_warnings() are not just single
word, except this "duplicate". Change it to " duplicate " would not
only fix for this test data, but also brings the consistency on
warning messages handling among the lvm commands.

Signed-off-by: Xiaoxue Wang <xiaoxwan@redhat.com>
Signed-off-by: max <mclinton@redhat.com>