chore: new 'data/insights_datasourcs' for datasource specs
All DatasourceProvider will be put in the new "insights_datasources"
directory, simple_command/foreach_execute/command_with_args which
depends on datasources are still stored in "insights_commands".

However, the data of "compliance/malware-detection" is still kept in
root directory as they are handled different with command insights
archive.

And datasources are renamed in this change.

Signed-off-by: Xiangce Liu <xiangceliu@redhat.com>