chore: new 'data/insights_datasources' for datasource specs (#4400)
Most of DatasourceProvider will be put in the new "insights_datasources"
directory

"simple_command/foreach_execute/command_with_args" specs which
depend on datasources are still stored in "insights_commands".
core built-in files, e.g. version_info, branch_info are still put in data root
the data of "compliance/malware-detection" is still kept in
data directory as they are handled different with command insights
archive.
And some datasources are renamed in this change.

Signed-off-by: Xiangce Liu <xiangceliu@redhat.com>