fix: support filters that start with minus sign (#4436)
The issue was the root cause of RHINENG-15287.

Commit d1065f5ec25ece259acd38c9cef78ac0721da863 resolved the issue using
reversed sorting, making a silent assumption that each component would
have at least one filter message that starts with a character with lower
precedence than `-`.

This commit uses the `grep` command line option `--` which causes all
following arguments to be interpreted as operands (positional arguments)
even if they start with `-`. See `info grep`. This works even if
a component has a single filter message and the filter message starts
with `-`.

Signed-off-by: Jan Holeček <71874510+jholecek-rh@users.noreply.github.com>