Show that the large_file_wf assertion is void
The assertion effectively checks that

```
len(broker[Stuff.large_file].content) < 101
```

The right part of the comparison is a complex way to compute the number 101.

When reading only 1024 bytes, the result contains only 51 lines. The
assertion passes even if no filters are applied (as documented by this
commit).

Similarly, there are 100 line in the sample file that contain `"9Some"`.
The assertion would pass even if the loaded data were not limited to
MAX_CONTENT_SIZE bytes.

For the given test settings, the assertions about `large_file_wf` do not
add anything to assertions about `large_file`.

Signed-off-by: Jan Holeček <71874510+jholecek-rh@users.noreply.github.com>