chore: deprecate LazyLogFileOutput (#4439)
The class was added based on the requirements of a downstream
project. It didn't meet the requirements of the project; the project
uses their own limited implementation now (no scans supported).

Specifically, LazyLogFileOutput is not a drop-in replacement for
LogFileOutput. It requires each and every rule to call
LazyLogFileOutput.do_scan() before using scans. This is error-prone;
if a rule forgets to do it, everything might work until the rule
is executed alone or until it is the first one executed.

LazyLogFileOutput should be deprecated and removed until we can invest
the effor into developing an implementation that will work as a full
drop-in replacement for LogFileOutput. We do not want downstream project
to become dependent on the current implementation.

Signed-off-by: Jan Holeček <71874510+jholecek-rh@users.noreply.github.com>
Signed-off-by: max <mclinton@redhat.com>