fix: Simplify error message for too long rule responses (#4477)
PR #4463 focused on improving the error message for the command line use
case - important details about the error were not available in the
command line.

However, insights-core is used also in services that are connected to
log analysis software. Log analysis software requires reasonably static
log messages for events aggregation to work properly. Specifically,
a log message containing the actual response size has proven to break
the aggregation because the response size can change in time even for
archives from a single system.

This commit is aiming at a compromise between the CLI and service use
cases within the constraints of the current logging setup used by the
insights-core:

- The response type and response key are included in the message because
  it is desirable to aggregate only events coming from the same rule.
- The size limit is included in the message to address the original
  issue that #4463 was solving. The size limit is expected to be stable
  enough in the service use cases not to compromise events aggregation.
- The actual rule response size is not included in the log message; it
  will not be available in the CLI use case for now. It is passed to log
  analysis software in the `extra` data.

A solution that would print log messages with variable data in the CLI
use case while not compromising events aggregation in log analysis
software is deferred to (a) future change(s).

Signed-off-by: Jan Holeček <71874510+jholecek-rh@users.noreply.github.com>