chore: enhance YAMLParser to parse yaml files containing = as value (#4508)
In must gather archive, there is a yaml file, it contains some lines like:

enum: 
- '!='
- =
- =~
- '!~'

Among them, - = line causes the following exception:
- could not determine a constructor for the tag 'tag:yaml.org,2002:value'

The root cause is: = is convert to yaml.org,2002:value, and for this tag, there is no constructor. To fix it, removing = from the yaml_implicit_resolvers, and making = is treated as a normal string.

Signed-off-by: Ping Qin <piqin@redhat.com>

rh-pre-commit.version: 2.3.2
rh-pre-commit.check-secrets: ENABLED