fix: use try_auto_configuration method for compliance (#4591)
The compliance datasource needs "legacy_upload=False", when
it's set as "legacy_upload=True". It's necessary to re-configure
the "base_url". Instead of processing the config.base_url
specially, using the "try_auto_configuration" provided by
client code, in which it processes the Satellite host content
in the standard way.

Jira: RHINENG-21116

Signed-off-by: Xiangce Liu <xiangceliu@redhat.com>